backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
openai
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
aiofiles
email_validator
